added testing and functionality to delete article, separated models and controllers and error handling
