+++ app.js
@@ -4,28 +4,20 @@ const cors = require("cors");
 const app = express();
 app.use(express.json());
 
+const {
+  customErrors,
+  psqlBasicError,
+  psqlComplexError,
+} = require("./error/error-handling");
+
 app.use(cors());
 app.use("/api", apiRouter);
 
-app.use((err, req, res, next) => {
-  if (err.status && err.msg) {
-    res.status(err.status).send({ msg: err.msg });
-  }
-  next(err);
-});
+app.use(customErrors);
 
-app.use((err, req, res, next) => {
-  if (err.code === "22P02") {
-    res.status(400).send({ msg: "Bad Request" });
-  }
-  next(err);
-});
+app.use(psqlBasicError);
+app.use(psqlComplexError);
 
-app.use((err, req, res, next) => {
-  if (err.code === "23503") {
-    res.status(404).send({ msg: "Path not found" });
-  }
-});
 app.all("*", (req, res) => {
   res.status(404).send({ msg: "Path not found" });
 });
