+++ model/users.js
@@ -1,0 +1,21 @@
+const db = require("../db/connection");
+
+exports.fetchUsers = () => {
+  return db.query("SELECT * FROM users;").then(({ rows }) => {
+    return rows;
+  });
+};
+
+exports.fetchUserByUsername = (username) => {
+  return db
+    .query("SELECT * FROM users WHERE users.username = $1;", [username])
+    .then(({ rows }) => {
+      if (rows.length === 0) {
+        return Promise.reject({
+          status: 404,
+          msg: "Path not found",
+        });
+      }
+      return rows[0];
+    });
+};
