+++ routes/articles-router.js
@@ -3,16 +3,21 @@ const {
   getArticles,
   getArticleById,
   patchArticle,
-  getComments,
-  postComment,
   postArticle,
-} = require("../controller/controller");
+  getComments,
+  deleteArticle,
+} = require("../controller/articles");
+const { postComment } = require("../controller/comments");
 
 articlesRouter.route("/").get(getArticles).post(postArticle);
 
-articlesRouter.route("/:article_id").get(getArticleById).patch(patchArticle);
-
 articlesRouter
+  .route("/:article_id")
+  .get(getArticleById)
+  .patch(patchArticle)
+  .delete(deleteArticle);
+
+delete articlesRouter
   .route("/:article_id/comments")
   .get(getComments)
   .post(postComment);
